Move common items to validateEmail package
