+++ app/src/main/java/com/realityexpander/tasky/di/AppModule.kt
@@ -5,6 +5,9 @@ import com.realityexpander.tasky.data.repository.remote.AuthApiFakeImpl
 import com.realityexpander.tasky.data.repository.local.AuthDaoFakeImpl
 import com.realityexpander.tasky.domain.IAuthRepository
 import com.realityexpander.tasky.domain.validation.*
+import com.realityexpander.tasky.domain.validation.validateEmail.EmailMatcherImpl
+import com.realityexpander.tasky.domain.validation.validateEmail.IValidateEmail
+import com.realityexpander.tasky.domain.validation.validateEmail.ValidateEmailImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
