+++ app/src/main/java/com/realityexpander/tasky/domain/validation/validateEmail/IEmailMatcher.kt
@@ -1,4 +1,4 @@
-package com.realityexpander.tasky.domain.validation
+package com.realityexpander.tasky.domain.validation.validateEmail
 
 interface IEmailMatcher {
     fun matches(email: String): Boolean
