+++ app/src/main/java/com/realityexpander/tasky/domain/validation/validateEmail/IValidateEmail.kt
@@ -1,4 +1,4 @@
-package com.realityexpander.tasky.domain.validation
+package com.realityexpander.tasky.domain.validation.validateEmail
 
 interface IValidateEmail {
     fun validate(email: String): Boolean
