+++ app/src/main/java/com/realityexpander/tasky/presentation/login_screen/LoginScreen.kt
@@ -39,7 +39,7 @@ import com.realityexpander.tasky.ui.util.UiText
 import com.realityexpander.tasky.data.repository.AuthRepositoryFakeImpl
 import com.realityexpander.tasky.data.repository.local.AuthDaoFakeImpl
 import com.realityexpander.tasky.data.repository.remote.AuthApiFakeImpl
-import com.realityexpander.tasky.domain.validation.ValidateEmailImpl
+import com.realityexpander.tasky.domain.validation.validateEmail.ValidateEmailImpl
 import com.realityexpander.tasky.domain.validation.ValidatePassword
 import com.realityexpander.tasky.domain.validation.ValidateUsername
 import com.realityexpander.tasky.presentation.destinations.RegisterScreenDestination
